+++ changelog.md
@@ -1,3 +1,10 @@
+# 11.1.0
+
+- Improve comments removal
+- Fix post list count in admin.
+- Disable XMLRPC
+- Disable unneeded APIs
+
 # 11.0.1
 
 - Fix featured images on page template